requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1725267488082/work
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1734858813237/work
distro==1.9.0
exceptiongroup==1.2.2
flake8==7.2.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1741793020411/work
flask-sock==0.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
iniconfig==2.1.0
isort==6.0.1
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
jiter==0.9.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
mccabe==0.7.0
mypy-extensions==1.0.0
openai==1.70.0
outcome==1.3.0.post0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1740663125313/work
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.13.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.2
pydantic_core==2.33.1
pyflakes==3.3.2
pyparsing==3.2.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytest==8.3.5
python-dotenv==1.1.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rsa==4.9
selenium==4.30.0
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wsproto==1.2.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard==0.23.0
instagrapi==2.1.2
